Moving from dhutils-lib to original worldedit classes Code anaylsis

+++ src/main/java/fr/free/jchecs/ai/NegaScoutEngine.java
@@ -64,7 +64,7 @@ final class NegaScoutEngine extends AbstractEngine
     final Integer resultat = TRANSPOSITIONS.get(pEtat, pProfondeur, pAlpha, pBeta);
     if (resultat != null)
     {
-      return resultat.intValue();
+        return resultat;
     }
 
     final boolean trait = pEtat.isWhiteActive();
